Improve android instrument tests
