+++ stream-video-android-core/src/androidTest/kotlin/io/getstream/video/android/core/RingTest.kt
@@ -73,6 +73,7 @@ class RingTest : IntegrationTestBase(connectCoordinatorWS = false) {
         assertSuccess(response)
         val acceptResponse = call.accept()
     }
+
     @Test
     fun reject() = runTest {
         val call = client.call("default")
