+++ stream-video-android-core/src/main/kotlin/io/getstream/video/android/core/MediaManager.kt
@@ -534,8 +534,12 @@ public class CameraManager(
 
     fun cleanup() {
         stopCapture()
-        videoCapturer.dispose()
-        surfaceTextureHelper.dispose()
+        if (::videoCapturer.isInitialized) {
+            videoCapturer.dispose()
+        }
+        if (::surfaceTextureHelper.isInitialized) {
+            surfaceTextureHelper.dispose()
+        }
         setupCompleted = false
     }
 }
